Fixed subtle De Bruijn index bug
This bug arises when you have a program like this:

\(a : *) -> \(a : a) -> \(a : a@1) -> a

The reason why is that the DeBruijn indices in the context are not updated
whenever you add new entries to the context.

The solution is to use `shift` to update all expressions stored within the
context when prepending a new bound variable.