Remove the `used` utility
Morte used to omit `Pi`-bound parameters if they were not used.  However, this
was problematic because it meant that rendering and parsing an expression would
transform all unused variable names to `_`.

At the time I originally wrote Morte, this was not such a bad thing, but as I
built the downstream `annah` library I realized that it's really worthwhile to
preserve these variable names since they are a really excellent form of built-in
documentation.  For example, when `annah` encodes the Bool type it generates
this expression:

    forall (Bool : *) -> forall (True : Bool) -> forall (True : Bool) -> Bool

... which is more self-descriptive and useful than this type:

    forall (Bool : *) -> Bool -> Bool -> Bool

This does mean that inferred Morte types will be more verbose, but I feel that
it's worth it overall.  I also think that it's particularly worthwhile for
Morte to preserve as much information as possible (including variable names)
considering its role as an intermediate language.