Switch from `happy` to `Earley` for parsing
There were two limitations of `happy` that motivated me to switch to another
parser.

The biggest limitation was that `happy` was not fully embedded within Haskell:
the parsing logic is specified in a separate language.  I ran into this
limitation when trying to write a custom parser for `ipfs`: the only way I
could provide a slightly different parser was to duplicate the entire parsing
file and tweak just a few lines.  Had the parser logic been fully embedded
within Haskell then I could have parametrized my parser on the part that
differed and avoided a lot of code duplication.

The second limitation was that `happy` only lets you return one parser
(typically for the entire expression).  If you want a parser for any other
sub-expression you need to duplicate the entire parsing code just to request
a separate parser.  These sorts of sub-expression parsers come in handy when
writing REPLs where you want to parse a fragment of the syntax and not
necessarily a complete expression.

The next question was what alternative parser to use.  I wanted to stick to an
ABNF-style parser so that `morte`'s parsing logic could be specified in a
language-agnostic manner.  I picked `Earley` because it was the only ABNF
parser I could find where 100% of the logic was completely embedded within the
Haskell language and without any Template Haskell.