Second attempt to fix test suite to run on `ghc-7.6.3`
The tests were failing due to the use of quasiquoters, which don't play well
with the `double-conversion` dependency because the quasiquoter will invoke
Template Haskell, which will in turn try to dynamically link in the
`double-conversion` and then fail (on older versions of `ghc`).

At first I thought that this `double-conversion` dependency was coming from the
quasiquoter package, but then I realized that it came from `morte` so the only
way to fix the tests would be to remove the quasiquoter altogether.

In order to remove the quasiquoter I had to factor out the test data into
separate data files that would be read in a test time, instead of inlining the
test programs within the test module.  This also required refactoring the
benchmark suite which also used data files so that the two suites wouldn't step
on each other's toes.