Change URLs to be stored as `Text` instead of `String`
I kept URLs as `String` for a long time for consistency with the API that
the `http-client` library expects for URLs, but this was causing huge
performance issues in the import logic due to the need to efficiently
manipulate the end of the string.  I switched the internals to use `Text`
everywhere for URLs now.